% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcofi.R
\name{calcofi_map}
\alias{calcofi_map}
\title{Map of CALCOFI sites}
\usage{
calcofi_map(
  geo =
    "https://raw.githubusercontent.com/marinebon/calcofi-analysis/master/data/plys_cinms.geojson",
  filter_str = "ply_code != \\"SoCal\\"",
  colors = c("red", "yellow")
)
}
\arguments{
\item{geo}{A geojson object that defines the polygons to be mapped.}

\item{filter_str}{A string used to filter in (or out) particular polygons.}

\item{colors}{A string vector that defines the colors of the mapped polygons.}
}
\value{
The output is an interactive map of CALCOFI sites overlaid on a coastal map of Southern California.
}
\description{
This function generates an interactive figure that shows the area in which
CALCOFI spring season net samples were located.
}
\examples{
\dontrun{
calcofi_map()
}

}
