% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocky.R
\name{rocky_tsplot}
\alias{rocky_tsplot}
\title{Plot time series of rocky intertidal monitoring data}
\usage{
rocky_tsplot(data, nms, sp_code, sp_name, sp_filter = "ALL")
}
\arguments{
\item{data}{Either \link{rocky_cover} or \link{rocky_counts}.}

\item{nms}{The National Marine Sanctuary code, e.g. "CINMS", from \link{sanctuaries}.}

\item{sp_code}{The species code of interest as referenced in \link{rocky_cover} or \link{rocky_counts} and described in \link{rocky_species}.}

\item{sp_name}{The species name used in the title of the plot.}

\item{sp_filter}{The filter value to apply for \link{rocky_cover} the \code{sp_target}
and for \link{rocky_counts} the \code{sp_method}. Defaults to "ALL".}
}
\value{
This function returns a dygraph object of the plotted time series data.
}
\description{
This function generates a plot of time series data collected by the Rocky Intertidal
Monitoring program.
}
\examples{
\dontrun{
rocky_tsplot(rocky_cover, "OCNMS", "CHTBAL", "Acorn Barnacles")
rocky_tsplot(rocky_cover, "CINMS", "CHTBAL", "Acorn Barnacles")
rocky_tsplot(rocky_counts, "OCNMS", "PISOCH", "Ochre Seastar")
}
}
