% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sanctuaries}
\alias{sanctuaries}
\title{Sanctuaries with spatial information}
\format{
A spatial feature tibble with the following columns:
\describe{
\item{nms}{national marine sanctuary acronym, e.g. CINMS}
\item{sanctuary}{name of National Marine Sanctuary, e.g. Channel Islands}
\item{url_zip}{link to zip file of original shapefile data}
\item{geom}{geometry of sanctuary}
}
}
\source{
\url{https://sanctuaries.noaa.gov/library/imast_gis.html}
}
\usage{
sanctuaries
}
\description{
A spatial features with sanctuary name and acronym. The default geometry is the unioned set of features found inside the spatial cell.
}
\keyword{datasets}
